% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{CustomApplicationExample-class}
\alias{CustomApplicationExample}
\alias{CustomApplicationExample-class}
\title{Example custom KBC application in R}
\description{
Example custom KBC application in R
}

